pdf-text-extractor/requirements.txt
pdfplumber
python-dotenv
tabulate